# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  updatedAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
}

export interface UserUpdateInput {
  /** @minLength 1 */
  name?: string;
  email?: string;
}

export interface UserResponse {
  data: User;
}

export interface UserListResponse {
  data: User[];
}