# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation failed
        "409":
          description: Duplicate unique value
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by id
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation failed
        "404":
          description: User not found
        "409":
          description: Duplicate unique value
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt
        - updatedAt
    UserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
      required:
        - name
        - email
    UserUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
      minProperties: 1
    UserResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"
      required:
        - data
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - data