# lib/api-zod/src/generated/types/userListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface UserListResponse {
  data: User[];
}